+++ src/main/java/com/compulsory/crudrestapi/controler/PlayerController.java
@@ -4,37 +4,39 @@ import com.compulsory.crudrestapi.model.Game;
 import com.compulsory.crudrestapi.model.Player;
 import com.compulsory.crudrestapi.repository.PlayerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Optional;
 
+@RestController
 public class PlayerController {
     @Autowired
     private PlayerRepository playerRepository;
 
     @PostMapping("/addPlayer")
-    public String saveGame(Player player) {
+    public String savePlayer(@RequestBody Player player) {
         playerRepository.save(player);
         return "Saved player with ID:" + player.getId();
     }
 
     @GetMapping("/findPlayers")
-    public List<Player> getAllGames() {
+    public List<Player> getAllPlayers() {
         return playerRepository.findAll();
     }
 
     @GetMapping("/findPlayers/{id}")
-    public Optional<Player> getGameByID(@PathVariable String id) {
+    public Optional<Player> getPlayerByID(@PathVariable String id) {
         return playerRepository.findById(Integer.parseInt(id));
     }
 
     @DeleteMapping("/deletePlayers/{id}")
-    public String deleteGameByID(@PathVariable String id) {
-        playerRepository.deleteById(Integer.parseInt(id));
-        return "Deleted player with ID:" + id;
+    public String deletePlayerByID(@PathVariable String id) {
+        if (!playerRepository.existsById(Integer.parseInt(id))) {
+            return "No player with this ID to be deleted.";
+        } else {
+            playerRepository.deleteById(Integer.parseInt(id));
+            return "Deleted player with ID:" + id;
+        }
     }
 }
